Moved to working entirely in float, no doubles for Eigen or anything using Eigen::VectorXf and Eigen::MatrixXf

+++ inc/PROCNP.h
@@ -37,11 +37,11 @@ namespace PROfit{
             float fixed_val;
 
             //Save last values for gradient calculation
-            Eigen::VectorXd last_param;
+            Eigen::VectorXf last_param;
             float last_value;
 
             bool correlated_systematics;
-            Eigen::MatrixXd prior_covariance;
+            Eigen::MatrixXf prior_covariance;
 
         public:
 
@@ -49,8 +49,8 @@ namespace PROfit{
             PROCNP(const std::string tag, const PROconfig *conin, const PROpeller *pin, const PROsyst *systin, const PROsc *oscin, const PROspec &datain, int nparams, int nsyst, EvalStrategy strat = EventByEvent, std::vector<float> physics_param_fixed = std::vector<float>());
 
             /*Function: operator() is what is passed to minimizer.*/
-            virtual double operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradient);
-            virtual double operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradient, bool nograd);
+            virtual float operator()(const Eigen::VectorXf &param, Eigen::VectorXf &gradient);
+            virtual float operator()(const Eigen::VectorXf &param, Eigen::VectorXf &gradient, bool nograd);
 
             PROmetric *Clone() const {
                 return new PROCNP(*this);
@@ -59,7 +59,7 @@ namespace PROfit{
             void reset() {
                 physics_param_fixed.clear();
                 last_value = 0;
-                last_param = Eigen::VectorXd::Constant(last_param.size(), 0);
+                last_param = Eigen::VectorXf::Constant(last_param.size(), 0);
             }
 
             void set_physics_param_fixed(const std::vector<float> &physics_param) {
@@ -73,9 +73,9 @@ namespace PROfit{
                 nparams += nsyst;
             }
 
-            float Pull(const Eigen::VectorXd &systs);
+            float Pull(const Eigen::VectorXf &systs);
 
-            void fixSpline(int fix, double valin);
+            void fixSpline(int fix, float valin);
 
             int nParams() const {return nparams;}
